Adding fix for double scrollbar in MacOS
Looks like it doesn't do the scrollbarSize calculation properly in macOS browsers (chrome, safari, firefox, etc) and ends up being 0. This seems to be happening only in MacOS. Windows and Linux does it well. So what I end up doing is validating for 0, if it's 0 it's almost certain that is under MacOS and it will add the corresponding 15px padding-right to hide the native scrollbar.